+++ src/Wolfgang.Etl.Transformers/PassThroughTransformer.cs
@@ -68,16 +68,7 @@ public sealed class PassThroughTransformer<T> : ITransformWithCancellationAsync<
     /// <exception cref="ArgumentNullException"><paramref name="items"/> is <see langword="null"/>.</exception>
     public IAsyncEnumerable<T> TransformAsync(IAsyncEnumerable<T> items)
     {
-#if NET6_0_OR_GREATER
         ArgumentNullException.ThrowIfNull(items);
-#else
-#pragma warning disable RCS1140 // Roslynator does not associate throw inside #else block with method XML doc
-        if (items == null)
-        {
-            throw new ArgumentNullException(nameof(items));
-        }
-#pragma warning restore RCS1140
-#endif
         return items;
     }
 
@@ -96,16 +87,7 @@ public sealed class PassThroughTransformer<T> : ITransformWithCancellationAsync<
     /// <exception cref="OperationCanceledException"><paramref name="token"/> was cancelled before or during enumeration.</exception>
     public IAsyncEnumerable<T> TransformAsync(IAsyncEnumerable<T> items, CancellationToken token)
     {
-#if NET6_0_OR_GREATER
         ArgumentNullException.ThrowIfNull(items);
-#else
-#pragma warning disable RCS1140 // Roslynator does not associate throw inside #else block with method XML doc
-        if (items == null)
-        {
-            throw new ArgumentNullException(nameof(items));
-        }
-#pragma warning restore RCS1140
-#endif
         return TransformAsyncCore(items, token);
     }